Crate mod_memoize

Structs

MemoizeParams

Memoized

Memoized is a helper type that allows native Rust types to be captured in memoization caches. Unfortunately, we cannot automatically make that work for all UserData that are exported to lua, but we can make it simple for a type to opt-in to that behavior.

Enums

CacheValue

MapKey

Functions

register